Fix average function to return double

+++ intlist.cpp
@@ -84,7 +84,7 @@ double IntList::average() const {
     if (count == 0) {
       return 0;
     }
-    return sum / count; // REPLACE THIS NON-SOLUTION
+    return (double)  sum / count; // REPLACE THIS NON-SOLUTION
 }
 
 // inserts value as new node at beginning of list
